requirements.txt
Click==7.0
cycler==0.10.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Script==2.0.6
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
livereload==2.5.0
MarkupSafe==1.1.1
matplotlib==3.0.3
numpy==1.16.3
pandas==0.24.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
scipy==1.2.1
six==1.12.0
tornado==6.0.2
visitor==0.1.3
Werkzeug==0.15.4
